Use yaml2obj for llvm-cov tests to see covmap
