+++ llvm/test/tools/llvm-cov/Inputs/branch-showBranchPercentage.yaml
@@ -1,0 +1,39 @@
+--- !ELF
+FileHeader:
+  Class:           ELFCLASS32
+  Data:            ELFDATA2LSB
+  Type:            ET_REL
+  Machine:         EM_ARM
+  Flags:           [ EF_ARM_EABI_VER5 ]
+  SectionHeaderStringTable: .strtab
+Sections:
+  - Name:            __llvm_covfun
+    Type:            SHT_PROGBITS
+    AddressAlign:    0x8
+    Content:         FAD58DE7366495DB70010000776050A22C20FE35CDAF69BBDD37045201001C0105021D1D2115190D110229292D0225023131352F050139374D45494145434D414967494145674D4145634D67494145674941453D513D5132010C012E02010207000F2005020007000F0500100183808080080501030304020304028380808008020203260202070700190200070024020007002E020008000E201D060008000E1D0012001820210A0012001815001D0024200E19001D00240D0028002E2012110028002E09002F0185808080080901050016020208000E2016290008000E2900120018201A2D0012001825001A009C8080800825001C001D1E00200023020208000E2031220008000E3100120018203526001200182A020F0014203902000F0014390016001939001A009B8080800839001B002D00060307938080800841010506132041330005000B670205041320453F0005000B630205021320494F0005000B5702050013204D630005000C3D00130383808080083D030308026F010602046F020C001420513D000C0014
+  - Name:            __llvm_covmap
+    Type:            SHT_PROGBITS
+    AddressAlign:    0x8
+    Content:         00000000260000000000000003000000012300222F746D702F6272616E63682D73686F774272616E636850657263656E746167652E630000
+  - Name:            __llvm_prf_names
+    Type:            SHT_PROGBITS
+    AddressAlign:    0x1
+    Content:         04006D61696E
+  - Type:            SectionHeaderTable
+    Sections:
+      - Name:            .strtab
+      - Name:            __llvm_covfun
+      - Name:            __llvm_covmap
+      - Name:            __llvm_prf_names
+      - Name:            .symtab
+Symbols:
+  - Name:            __llvm_prf_names
+    Type:            STT_SECTION
+    Section:         __llvm_prf_names
+  - Name:            __covrec_DB956436E78DD5FAu
+    Type:            STT_OBJECT
+    Section:         __llvm_covfun
+    Binding:         STB_WEAK
+    Size:            0x18C
+    Other:           [ STV_HIDDEN ]
+...
